Ordering Microservice: cancel slice command

+++ Services/Ordering/Biz.BrightOnion.Ordering.Domain/AggregatesModel/OrderAggregate/IOrderRepository.cs
@@ -10,7 +10,7 @@ namespace Biz.BrightOnion.Ordering.Domain.AggregatesModel.OrderAggregate
 
     void Update(Order order);
 
-    Task<Order> GetAsync(int orderId);
+    Task<Order> GetAsync(long orderId);
 
     Task<Order> GetByDayAndRoomEagerAsync(DateTime day, long roomId);
   }
